+++ lib/cupertino_native.dart
@@ -10,6 +10,7 @@ export 'components/tab_bar.dart';
 export 'components/popup_menu_button.dart';
 export 'style/sf_symbol.dart';
 export 'style/button_style.dart';
+export 'style/image_placement.dart';
 export 'components/button.dart';
 
 import 'cupertino_native_platform_interface.dart';